+++ haruka/modules/haruka_info.py
@@ -55,7 +55,7 @@ class HarukaInfoMod(loader.Module):
             ),
             loader.ConfigValue(
                 "banner_url",
-                "",
+                None,
                 lambda: self.strings("_cfg_banner"),
                 validator=loader.validators.RandomLink(),
             ),